Mv vm types to separate file
